+++ es5/OBSWebSocket.js
@@ -12,7 +12,6 @@ require('./util/callbackPromise')(Promise);
 var Socket = require('./Socket');
 var Status = require('./Status');
 var debug = require('debug')('obs-websocket-js:Core');
-var API = require('./API');
 
 var requestCounter = 0;
 
@@ -26,33 +25,22 @@ var OBSWebSocket = function (_Socket) {
   function OBSWebSocket() {
     _classCallCheck(this, OBSWebSocket);
 
-    var _this = _possibleConstructorReturn(this, (OBSWebSocket.__proto__ || Object.getPrototypeOf(OBSWebSocket)).call(this));
-
-    _this.availableRequests = [];
-    _this.availableEvents = [];
-
-    _this.registerRequest(API.availableMethods);
-    _this.registerEvent(API.availableEvents);
-
-    _this.registerRequest(['GetAuthRequired', 'Authenticate']);
-    _this.registerEvent(['ConnectionOpened', 'ConnectionClosed', 'AuthenticationSuccess', 'AuthenticationFailure']);
-    return _this;
+    return _possibleConstructorReturn(this, (OBSWebSocket.__proto__ || Object.getPrototypeOf(OBSWebSocket)).apply(this, arguments));
   }
-
-  /**
-   * Internal generic Socket request method. Returns a promise, handles callbacks.
-   * Generates a messageId internally and will override any passed in the args.
-   * Note that the requestType here is pre-marshaling and currently must match exactly what the websocket plugin is expecting.
-   *
-   * @param  {String}   requestType obs-websocket plugin expected request type.
-   * @param  {Object}   [args={}]   request arguments.
-   * @param  {Function} callback    Optional. callback(err, data)
-   * @return {Promise}              Promise, passes the plugin response object.
-   */
-
 
   _createClass(OBSWebSocket, [{
     key: 'send',
+
+    /**
+     * Internal generic Socket request method. Returns a promise, handles callbacks.
+     * Generates a messageId internally and will override any passed in the args.
+     * Note that the requestType here is pre-marshaling and currently must match exactly what the websocket plugin is expecting.
+     *
+     * @param  {String}   requestType obs-websocket plugin expected request type.
+     * @param  {Object}   [args={}]   request arguments.
+     * @param  {Function} callback    Optional. callback(err, data)
+     * @return {Promise}              Promise, passes the plugin response object.
+     */
     value: function send(requestType) {
       var _this2 = this;
 
@@ -104,70 +92,6 @@ var OBSWebSocket = function (_Socket) {
           _this2.emit(`obs:internal:message:id-${messageId}`, rejectReason);
         }
       }).callback(callback);
-    }
-
-    /**
-     * Add a new recognized request.
-     * Enables usage with the following syntaxes.
-     * `obs.requestName({args}, callback(err, data)) returns Promise`
-     * `obs.RequestName({args}, callback(err, data)) returns Promise`
-     *
-     * @param  {Array}  [requestNames=[]] String or Array of String request names as defined by the obs-websocket plugin.
-     */
-
-  }, {
-    key: 'registerRequest',
-    value: function registerRequest() {
-      var _this3 = this;
-
-      var requestNames = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : [];
-
-      if (!Array.isArray(requestNames)) {
-        requestNames = [requestNames];
-      }
-
-      requestNames.forEach(function (requestName) {
-        _this3.availableRequests.push(requestName);
-        var handler = function handler(args, callback) {
-          return this.send(requestName, args, callback);
-        };
-
-        _this3[requestName] = handler;
-        _this3[requestName.charAt(0).toLowerCase() + requestName.slice(1)] = handler;
-      });
-    }
-
-    /**
-     * Add a new recognized event.
-     * Enables usage with the following syntax.
-     * `obs.onEventName(callback(data))`
-     *
-     * @param  {Array}  [eventNames=[]] String or Array of String event names as defined by the obs-websocket plugin.
-     */
-
-  }, {
-    key: 'registerEvent',
-    value: function registerEvent() {
-      var _this4 = this;
-
-      var eventNames = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : [];
-
-      if (!Array.isArray(eventNames)) {
-        eventNames = [eventNames];
-      }
-
-      eventNames.forEach(function (eventName) {
-        _this4.availableEvents.push(eventName);
-        _this4[`on${eventName}`] = function (callback) {
-          if (typeof callback !== 'function') {
-            return;
-          }
-
-          this.on(eventName, function (data) {
-            callback(data);
-          });
-        };
-      });
     }
   }]);