+++ test/requests.spec.js
@@ -26,18 +26,6 @@ test('rejects when an invalid request is sent', async t => {
   t.deepEqual(resp.error, 'invalid request type');
 });
 
-test('allows using lowercase request methods', async t => {
-  obs.registerRequest('ValidMethodName');
-  const resp = await obs.validMethodName();
-  t.deepEqual(resp.status, 'ok');
-});
-
-test('allows using uppercase request methods', async t => {
-  obs.registerRequest('ValidMethodName');
-  const resp = await obs.ValidMethodName(); // eslint-disable-line new-cap
-  t.deepEqual(resp.status, 'ok');
-});
-
 test('permits null args', async t => {
   await t.notThrows(obs.send('ValidMethodName', null));
 });
@@ -58,12 +46,6 @@ test.cb('allows the use of a callback instead of a promise.reject', t => {
     t.deepEqual(err.status, 'error');
     t.end();
   });
-});
-
-test('assigns default methods based on API.js', async t => {
-  const resp = await obs.getAuthRequired();
-  t.deepEqual(resp.status, 'ok');
-  t.false(resp.authRequired);
 });
 
 test('rejects when no open connection exists', async t => {